gctrace: better GCs/sec metric
Currently GCs/sec uses the end time of the last GC as the duration.
This has several problems: forced GCs at the beginning tend to bring
this down, and if there's a long time between the last GC and the
program exiting (an easy problem to have if the heap is growing
rapidly or getting very large), this tends to be higher than it should
be.

Fix this by instead using the time between the first non-forced GC and
the end of execution.