gcbench: fill histogram width
Currently the latency histogram is sized to fit in the specified width
while combining integer multiples of underlying buckets. This results
in annoying shifts in actual histogram width in response to minor
changes in the data.

Fix this by resampling the underlying buckets into the histogram. We
use an algorithm that's sufficiently stable I wasn't able to construct
a histogram whose resampled total didn't equal that of the underlying
buckets.